:recycle:Refactor 컨트롤러 SVG 변환 서비스 일원화
